[6.14.z Cherrypick PR#13450] Move Registration test cases
